backend/requirements.txt
asgiref==3.5.2
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1;python_version<"3.9"
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
distro==1.4.0
Django==3.2.13
django-cors-headers==3.13.0
djangorestframework==3.13.1
entrypoints==0.3
gevent
greenlet
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces
oauthlib==3.1.0
pbr==5.9.0
pexpect==4.6.0
psycopg2-binary
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyHamcrest==1.9.0
pyinotify==0.9.6
python-decouple
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
pytz==2022.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
ssh-import-id==5.10
testresources==2.0.1
Twisted
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.event==4.5.0
zope.interface
